requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
anime-downloader==5.0.9
APScheduler==3.6.3
async-lru==1.0.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
better-profanity==0.7.0
bs4==0.0.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
cfscrape==2.1.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coloredlogs==15.0
cryptography==3.4.7
cssselect2==0.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.1.0
emoji==1.2.0
fake-useragent==0.1.11
Faker==8.1.3
feedparser==6.0.2
ffmpeg-python==0.2.0
filesplit==3.0.2
future==0.18.2
fuzzywuzzy==0.18.0
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.7
GitPython==3.1.14
glitch-this==1.0.2
google-trans-new==1.1.9
googletrans==3.0.0
greenlet==1.1.0
gTTS==2.2.2
h11==0.9.0
h2==3.2.0
hachoir==3.1.2
heroku3==4.2.3
hpack==3.0.0
hstspreload==2020.12.22
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==9.1
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
img2pdf==0.4.1
iniconfig==1.1.1
itsdangerous==1.1.0
kiwisolver==1.3.1
langdetect==1.0.9
lottie==0.6.7
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.2
motor==2.4.0
multidict==5.1.0
mysql-connector==2.2.9
mysqlclient==2.0.3
networkx==2.5.1
NoteShrinker==0.2.0
numpy==1.20.2
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
packaging==20.9
pikepdf==2.12.0
Pillow==8.2.0
pluggy==0.13.1
psutil==5.8.0
psycopg2-binary==2.8.6
py==1.10.0
pyaes==1.6.1
pycparser==2.20
pycryptodome==3.10.1
pygifsicle==1.0.4
pymediainfo==5.1.0
pymongo==3.11.4
pyparsing==2.4.7
pyreadline==2.1
Pyrogram==1.2.9
pyshorteners==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
pytesseract==0.3.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
pyvirtualcam==0.8.0
PyWavelets==1.1.1
requests==2.25.1
requests-cache==0.6.3
rfc3986==1.5.0
scikit-image==0.18.1
scipy==1.6.3
selenium==3.141.0
sgmllib3k==1.0.0
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
tabulate==0.8.9
telegraph==1.4.1
text-unidecode==1.3
TgCrypto==1.2.2
tifffile==2021.4.8
tinycss2==1.1.0
tinydb==4.4.0
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
url-normalize==1.4.3
urllib3==1.26.4
webencodings==0.5.1
wget==3.2
yarl==1.6.3
youtube-dl==2021.4.26
youtube-search==2.1.0
youtube-search-python==1.4.5
